Move color manipulation code to colors pacakge
