+++ white_brush/colors/color_extraction.py
